abstracted realtime elements in activity card to seperate components (#132)
